git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/lno-branch@79771 138bc75d-0d04-0410-961f-82ee72b054a4

+++ libbanshee/engine/util.c
@@ -172,9 +172,9 @@ char *ptr_to_ascii(void *ptr) {
 }
 
 /* Convert a pointer to an integer */
-int ptr_hash(void *ptr)
+long ptr_hash(void *ptr)
 {
-  return (int) ptr;
+  return (long) ptr;
 }
 
 /* Return TRUE iff ptr1 == ptr2 */
